chore: Moving some functions to records context
